Inline Box demo configuration
